Clean up READMEs and fix examples
