--- examples/basic/mcp_agent_server/mcp_agent.config.yaml
@@ -1,34 +1,0 @@
-$schema: ../../../schema/mcp-agent.config.schema.json
-
-execution_engine: asyncio
-logger:
-  type: file
-  level: debug
-  batch_size: 100
-  flush_interval: 2
-  max_queue_size: 2048
-  http_endpoint:
-  http_headers:
-  http_timeout: 5
-
-mcp:
-  servers:
-    mcp_agent:
-      command: "uv"
-      args: ["run", "server.py"]
-    fetch:
-      command: "uvx"
-      args: ["mcp-server-fetch"]
-    filesystem:
-      command: "npx"
-      args:
-        [
-          "-y",
-          "@modelcontextprotocol/server-filesystem",
-          "/Users/saqadri/Desktop",
-          "/Users/saqadri/Downloads",
-        ]
-
-openai:
-  # Secrets (API keys, etc.) are stored in an mcp_agent.secrets.yaml file which can be gitignored
-  default_model: gpt-4o
