--- examples/basic/mcp_agent_server/mcp_agent.secrets.yaml.example
@@ -1,7 +1,0 @@
-$schema: ../../../schema/mcp-agent.config.schema.json
-
-openai:
-  api_key: openai_api_key
-
-anthropic:
-  api_key: anthropic_api_key
