--- examples/temporal_orchestrator_example/graded_report.md
@@ -1,77 +1,0 @@
-# Graded Report for "The Battle of Glimmerwood"
-
-## Proofreading Feedback
-
-### Grammar, Spelling, and Punctuation:
-
-1. **Punctuation with Introductory Phrases:**
-   - **Original Sentence:** Amidst the chaos, a young girl named Elara stood her ground; she rallied the villagers and devised a clever plan.
-   - **Revised Suggestion:** Amidst the chaos, a young girl named Elara stood her ground. She rallied the villagers and devised a clever plan.
-   - **Feedback:** The semicolon can be replaced with a period to make two complete sentences for better readability.
-
-2. **Comma Usage:**
-   - **Original Sentence:** Using the forest's natural defenses, Elara and the villagers lured the marauders into a trap.
-   - **Revised Suggestion:** Correct as is.
-   - **Feedback:** The comma after the introductory phrase is correct and necessary.
-
-3. **Compound Sentence Structure:**
-   - **Original Sentence:** As the bandits approached the village square, a herd of Glimmerfoxes emerged, blinding the marauders with their dazzling light, and the villagers seized the opportunity to capture the invaders.
-   - **Revised Suggestion:** As the bandits approached the village square, a herd of Glimmerfoxes emerged, blinding the marauders with their dazzling light. The villagers seized the opportunity to capture the invaders.
-   - **Feedback:** Breaking the sentence into two separate sentences can improve clarity.
-
-4. **Use of Semicolon:**
-   - **Original Sentence:** Elara stood her ground; she rallied the villagers and devised a clever plan.
-   - **Revised Suggestion:** Elara stood her ground. She rallied the villagers and devised a clever plan.
-   - **Feedback:** While the semicolon is grammatically correct, using a period can enhance the narrative's impact by providing a pause.
-
-5. **Subject-Verb Agreement:**
-   - **Original Sentence:** The villagers, who lived peacefully, shared their home with the forest's magical creatures, especially the Glimmerfoxes, whose fur shimmered like moonlight.
-   - **Revised Suggestion:** Correct as is.
-   - **Feedback:** The sentence is grammatically correct with proper subject-verb agreement.
-
-6. **Clarity in Expression:**
-   - **Original Sentence:** However, not everything was as it seemed.
-   - **Revised Suggestion:** Correct as is.
-   - **Feedback:** The sentence effectively introduces suspense; no changes needed.
-
-#### Overall Feedback:
-
-The story is well-written with minor suggestions for punctuation improvements to enhance clarity and readability. It maintains grammatical correctness, and there are no spelling errors. Adjustments for sentence structure and punctuation will enhance clarity and readability while maintaining the engaging tone.
-
-## Factual and Logical Consistency
-
-1. **Setting Consistency:**
-   - The mystical setting with radiant trees and magical creatures like Glimmerfoxes is consistent throughout, forming a solid fantasy backdrop.
-
-2. **Character Actions and Responses:**
-   - The protagonist Elara is consistent in her leadership role, leveraging the forest’s defenses against the marauders.
-
-3. **Logical Consistency:**
-   - The narrative sequence is coherent, though the unresolved mystery of Glimmerstones adds an open-ended feel to the story.
-
-There are no contradictions, and the magical elements are effective and consistent throughout.
-
-## Style Adherence and Narrative Analysis
-
-### Narrative Flow
-- The story follows a clear structure with engaging pacing and vivid imagery, maintaining coherence and reader interest.
-
-### Tone
-- The tone is consistent with a sense of adventure and mysticism, appropriate for the plot.
-
-### Suggestions for Improvement
-- **Expand on Consequences:** Delve deeper into the Glimmerstones' mythology and implications for Elara.
-- **Character Development:** Enrich Elara’s background for stronger reader connection.
-- **Tension Building:** Introduce more tension before the climax, enhancing reader excitement.
-
-## Updated Feedback on APA Reference Style Adherence
-
-### References and Citations
-- **Structure and Consistency:** Consider in-text citations for factual claims and a "References" section following APA guidelines.
-  
-### Suggestions for Enhancement
-- Ensure any factual elements or specific quotations in the story are properly cited in APA format, with consistency in document formatting for clarity and readability.
-
----
-
-With these insights, the story "The Battle of Glimmerwood" can be further refined for clarity, depth, and adherence to stylistic standards.
