--- examples/temporal_orchestrator_example/requirements.txt
@@ -1,5 +1,0 @@
-# Core framework dependency
-mcp-agent @ file://../../  # Link to the local mcp-agent project root
-
-# Additional dependencies specific to this example
-temporalio
