--- examples/temporal_orchestrator_example/short_story.md
@@ -1,11 +1,0 @@
-## The Battle of Glimmerwood
-
-In the heart of Glimmerwood, a mystical forest known for its radiant trees, a small village thrived. The villagers, who lived peacefully, shared their home with the forest's magical creatures, especially the Glimmerfoxes, whose fur shimmered like moonlight.
-
-One fateful evening, the peace was shattered when the infamous Dark Marauders attacked. Led by the cunning Captain Thorn, the bandits aimed to steal the precious Glimmerstones, which were believed to grant immortality.
-
-Amidst the chaos, a young girl named Elara stood her ground; she rallied the villagers and devised a clever plan. Using the forest's natural defenses, Elara and the villagers lured the marauders into a trap. As the bandits approached the village square, a herd of Glimmerfoxes emerged, blinding the marauders with their dazzling light, and the villagers seized the opportunity to capture the invaders.
-
-Elara's bravery was celebrated, and she was hailed as the Guardian of Glimmerwood. The Glimmerstones were secured in a hidden grove protected by an ancient spell.
-
-However, not everything was as it seemed. The true power of the Glimmerstones was never confirmed, and whispers of a hidden agenda lingered among the villagers.
