--- examples/workflow_mcp_server/mcp_agent.secrets.yaml.example
@@ -1,5 +1,0 @@
-openai:
-  api_key: sk-your-openai-key
-
-anthropic:
-  api_key: sk-ant-your-anthropic-key
